chore: Unify GapJumpingController and StallDetector in the same file (#8148)
This helps debugging issues easier by having everything in the same
file. It also allows you to align your implementation with what hls.js
and dash.js do.
